modified for int64
	modified:   paddle/fluid/pir/drr/src/ir_operation_factory.cc

+++ paddle/fluid/pir/drr/src/ir_operation_factory.cc
@@ -433,7 +433,7 @@ void OperationFactory::RegisterManualOpCreator() {
                             common::errors::InvalidArgument(
                                 "'dilations' Attribute is expected for "
                                 "Pool2dOp. "));
-          std::vector<int> dilations;
+          std::vector<int64_t> dilations;
           for (size_t i = 0;
                i < attrs.at("dilations").dyn_cast<pir::ArrayAttribute>().size();
                i++) {
